refactor(edit): split pipeline edit batch codec into JSON, validate, and encode modules
Keep PipelineEditBatch as the envelope facade and follow the NM6P named-module pattern so nested JSON, validate, and encode/decode are separate modules rather than a method-file split.
